+++ R/xrect.R
@@ -1,4 +1,4 @@
-#' xrect
+#' Rectangles
 #'
 #' Draw rectangles from four colummns xmin,xmax,ymin,ymax
 #'
@@ -17,7 +17,6 @@
 #' xrect(sort(runif(100)))
 #' xrect(runif(100), col = hcl.colors(25, alpha = seq(.2, .8, length.out = 25)))
 #'
-#' #example(xcontour, ask = FALSE)
 #' ex <- c(0.2, 0.8, .2, .6)
 #' xrect(ex, add = TRUE, lwd = 5, lty = 2)
 xrect <- function(x, add = FALSE, ..., asp = 1L) {